+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Reduce Dougie addressing user by name every message
 * Fix mobile chat textarea height to match buttons dynamically
 * Dougie sees message reactions and who reacted
+* Dougie reads today's spending and recent transactions from local DB for real-time data
 
 ### 1.11.0: 2026-03-27